requirements.txt
Flask==3.1.0
gunicorn==23.0.0
PyPDF2==3.0.1
python-docx==1.1.2
scikit-learn==1.6.1
